feat(tooling): Add netlify to publish for testing
tooling: split wdio config into separate files
