+++ .eslintignore
@@ -1,3 +1,4 @@
 **/dist/**
 **/es/**
 **/cjs/**
+dist-test/**
