+++ package.json
@@ -11,20 +11,20 @@
     "build:widgets": "cross-env NODE_ENV=${NODE_ENV:-production} npm run build widgets",
     "build:sri": "npm run build sri",
     "clean": "npm run clean:dist && npm run clean:transpile",
-    "clean:dist": "rimraf \"packages/**/dist/\"",
+    "clean:dist": "rimraf \"packages/**/dist/\" && rimraf \"packages/**/dist/\"",
     "clean:transpile": "rimraf \"packages/**/es/\" && rimraf \"packages/**/cjs/\"",
     "test": "npm run static-analysis && npm run jest",
     "test:clean": "npm run clean && npm run test",
-    "test:automation": "wdio wdio.conf.js",
-    "test:automation:oneOnOne": "wdio wdio.conf.js --suite oneOnOne",
-    "test:automation:space": "wdio wdio.conf.js --suite space",
-    "test:automation:recents": "wdio wdio.conf.js --suite recents",
-    "test:tap": "cross-env TAP=true wdio wdio.conf.js --suite tap",
-    "test:integration": "cross-env INTEGRATION=true wdio wdio.conf.js --suite integration",
+    "test:automation": "wdio ./wdio/wdio.conf.js",
+    "test:automation:oneOnOne": "wdio ./wdio/wdio.conf.js --suite oneOnOne",
+    "test:automation:space": "wdio ./wdio/wdio.conf.js --suite space",
+    "test:automation:recents": "wdio ./wdio/wdio.conf.js --suite recents",
+    "test:tap": "wdio ./wdio/wdio.tap.conf.js --suite tap",
+    "test:integration": "cross-env INTEGRATION=true wdio ./wdio/wdio.conf.js --suite integration",
     "jest": "cross-env TZ=utc jest --config=jest.config.json",
     "static-analysis": "npm run eslint",
     "lint:eslint": "eslint --ignore-path .gitignore $(npm run --silent lint:eslint:ci-reporter)",
-    "eslint": "eslint ./",
+    "eslint": "eslint ./src ./scripts ./packages",
     "serve": "npm run start",
     "stylelint": "stylelint \"./packages/node_modules/**/*.css\"",
     "start": "cross-env NODE_ENV=${NODE_ENV:-development} ./scripts/start/index.js",
@@ -156,6 +156,7 @@
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
     "mockdate": "^2.0.1",
+    "netlify-cli": "^1.2.2",
     "npm-check-updates": "^2.12.1",
     "postcss": "^6.0.3",
     "postcss-cli": "^5.0.0",
