+++ wdio/wdio.tap.conf.js
@@ -1,0 +1,8 @@
+require('dotenv').config();
+
+const config = require('./wdio.sauce.conf.js');
+
+config.baseUrl = 'https://code.s4d.io';
+config.sauceConnect = false;
+
+exports.config = config;
